split actions to files
