+++ src/actions/fetchApi.js
@@ -1,0 +1,5 @@
+import 'whatwg-fetch';
+
+const fetchApi = url => fetch(`/api/${url}`).then(response => response.json());
+
+export default fetchApi;
