--- src/constants/actionTypes.js
@@ -1,6 +1,0 @@
-export const REQUEST_WEEKNUMBER = 'REQUEST_WEEKNUMBER';
-export const RECEIVE_WEEKNUMBER = 'RECEIVE_WEEKNUMBER';
-export const REQUEST_BUGSDIFF = 'REQUEST_BUGSDIFF';
-export const RECEIVE_BUGSDIFF = 'RECEIVE_BUGSDIFF';
-export const REQUEST_BUGSHISTORY = 'REQUEST_BUGSHISTORY';
-export const RECEIVE_BUGSHISTORY = 'RECEIVE_BUGSHISTORY';
